Consolidate mobile WebView editor logic into shared packages/ui
Extract duplicated CRDT lifecycle, node views, record layout, and database
layout from the mobile WebView editor into shared primitives in packages/ui.
