+++ apps/mobile/webviews/editor/src/extensions/folder.tsx
@@ -1,27 +1,8 @@
-import { mergeAttributes, Node } from '@tiptap/core';
-import { ReactNodeViewRenderer } from '@tiptap/react';
+import { FolderNodeView } from '@colanode/ui/editor/views';
+import { createNodeExtension } from '@colanode/ui/editor/extensions/create-node-extension';
 
-import { MobileFolderNodeView } from '../views/folder';
-
-export const MobileFolderNode = Node.create({
+export const MobileFolderNode = createNodeExtension({
   name: 'folder',
-  group: 'block',
-  atom: true,
-  defining: true,
   draggable: false,
-  addAttributes() {
-    return {
-      id: {
-        default: null,
-      },
-    };
-  },
-  renderHTML({ HTMLAttributes }) {
-    return ['folder', mergeAttributes(HTMLAttributes)];
-  },
-  addNodeView() {
-    return ReactNodeViewRenderer(MobileFolderNodeView, {
-      as: 'folder',
-    });
-  },
+  view: FolderNodeView,
 });
