+++ packages/ui/src/components/records/record-container.tsx
@@ -1,10 +1,7 @@
 import { LocalRecordNode } from '@colanode/client/types';
 import { NodeRole, hasNodeRole } from '@colanode/core';
 import { Document } from '@colanode/ui/components/documents/document';
-import { RecordAttributes } from '@colanode/ui/components/records/record-attributes';
-import { RecordDatabase } from '@colanode/ui/components/records/record-database';
-import { RecordProvider } from '@colanode/ui/components/records/record-provider';
-import { Separator } from '@colanode/ui/components/ui/separator';
+import { RecordDetailLayout } from '@colanode/ui/components/records/record-detail-layout';
 import { useWorkspace } from '@colanode/ui/contexts/workspace';
 
 interface RecordContainerProps {
@@ -18,12 +15,8 @@ export const RecordContainer = ({ record, role }: RecordContainerProps) => {
   const canEdit =
     record.createdBy === workspace.userId || hasNodeRole(role, 'editor');
   return (
-    <RecordDatabase id={record.databaseId} role={role}>
-      <RecordProvider record={record} role={role}>
-        <RecordAttributes />
-      </RecordProvider>
-      <Separator className="my-4 w-full" />
+    <RecordDetailLayout record={record} role={role}>
       <Document node={record} canEdit={canEdit} autoFocus={false} />
-    </RecordDatabase>
+    </RecordDetailLayout>
   );
 };
