+++ packages/ui/src/editor/extensions/folder.tsx
@@ -1,27 +1,8 @@
-import { mergeAttributes, Node } from '@tiptap/core';
-import { ReactNodeViewRenderer } from '@tiptap/react';
-
 import { FolderNodeView } from '@colanode/ui/editor/views';
+import { createNodeExtension } from '@colanode/ui/editor/extensions/create-node-extension';
 
-export const FolderNode = Node.create({
+export const FolderNode = createNodeExtension({
   name: 'folder',
-  group: 'block',
-  atom: true,
-  defining: true,
   draggable: true,
-  addAttributes() {
-    return {
-      id: {
-        default: null,
-      },
-    };
-  },
-  renderHTML({ HTMLAttributes }) {
-    return ['folder', mergeAttributes(HTMLAttributes)];
-  },
-  addNodeView() {
-    return ReactNodeViewRenderer(FolderNodeView, {
-      as: 'folder',
-    });
-  },
+  view: FolderNodeView,
 });
