+++ packages/ui/src/editor/views/index.tsx
@@ -3,11 +3,14 @@ import { DatabaseNodeView } from '@colanode/ui/editor/views/database';
 import { FileNodeView } from '@colanode/ui/editor/views/file';
 import { FolderNodeView } from '@colanode/ui/editor/views/folder';
 import { MentionNodeView } from '@colanode/ui/editor/views/mention';
+import { NodeLink } from '@colanode/ui/editor/views/node-link';
 import { PageNodeView } from '@colanode/ui/editor/views/page';
 import { TableNodeView } from '@colanode/ui/editor/views/table';
 import { TableCellNodeView } from '@colanode/ui/editor/views/table-cell';
 import { TableHeaderNodeView } from '@colanode/ui/editor/views/table-header';
 import { TempFileNodeView } from '@colanode/ui/editor/views/temp-file';
+import { useReferencedNode } from '@colanode/ui/editor/views/use-referenced-node';
+import { useReferencedUser } from '@colanode/ui/editor/views/use-referenced-user';
 
 export {
   CodeBlockNodeView,
@@ -16,8 +19,11 @@ export {
   TempFileNodeView,
   FolderNodeView,
   MentionNodeView,
+  NodeLink,
   PageNodeView,
   TableCellNodeView,
   TableNodeView,
   TableHeaderNodeView,
+  useReferencedNode,
+  useReferencedUser,
 };
