+++ packages/ui/src/editor/views/mention.tsx
@@ -1,30 +1,15 @@
-import { eq, useLiveQuery } from '@tanstack/react-db';
 import { type NodeViewProps } from '@tiptap/core';
 import { NodeViewWrapper } from '@tiptap/react';
 
 import { Avatar } from '@colanode/ui/components/avatars/avatar';
-import { useWorkspace } from '@colanode/ui/contexts/workspace';
 import { defaultClasses } from '@colanode/ui/editor/classes';
+import { useReferencedUser } from '@colanode/ui/editor/views/use-referenced-user';
 
 export const MentionNodeView = ({ node }: NodeViewProps) => {
-  const workspace = useWorkspace();
-
   const target = node.attrs.target;
-  const userQuery = useLiveQuery(
-    (q) =>
-      q
-        .from({ users: workspace.collections.users })
-        .where(({ users }) => eq(users.id, target))
-        .select(({ users }) => ({
-          id: users.id,
-          name: users.name,
-          avatar: users.avatar,
-        }))
-        .findOne(),
-    [workspace.userId, target]
-  );
 
-  const user = userQuery.data;
+  const { user } = useReferencedUser(target);
+
   const name = user?.name ?? 'Unknown';
   const avatar = user?.avatar;
