+++ packages/ui/src/editor/views/page.tsx
@@ -1,40 +1,17 @@
-import { eq, useLiveQuery } from '@tanstack/react-db';
 import { type NodeViewProps } from '@tiptap/core';
 import { NodeViewWrapper } from '@tiptap/react';
 
 import { LocalPageNode } from '@colanode/client/types';
 import { Avatar } from '@colanode/ui/components/avatars/avatar';
-import { Link } from '@colanode/ui/components/ui/link';
-import { useWorkspace } from '@colanode/ui/contexts/workspace';
+import { NodeLink } from '@colanode/ui/editor/views/node-link';
+import { useReferencedNode } from '@colanode/ui/editor/views/use-referenced-node';
 
 export const PageNodeView = ({ node }: NodeViewProps) => {
-  const workspace = useWorkspace();
-
   const id = node.attrs.id;
 
-  if (!id) {
-    return null;
-  }
+  const { node: page, isLoading } = useReferencedNode<LocalPageNode>(id);
 
-  const pageGetQuery = useLiveQuery(
-    (q) =>
-      q
-        .from({ nodes: workspace.collections.nodes })
-        .where(({ nodes }) => eq(nodes.id, id))
-        .findOne(),
-    [workspace.userId, id]
-  );
-
-  if (
-    pageGetQuery.isLoading ||
-    !pageGetQuery.data ||
-    pageGetQuery.data.type !== 'page'
-  ) {
-    return null;
-  }
-
-  const page = pageGetQuery.data as LocalPageNode | undefined;
-  if (!page) {
+  if (!id || isLoading || !page || page.type !== 'page') {
     return null;
   }
 
@@ -43,14 +20,14 @@ export const PageNodeView = ({ node }: NodeViewProps) => {
 
   return (
     <NodeViewWrapper data-id={node.attrs.id}>
-      <Link from="/workspace/$userId" to="$nodeId" params={{ nodeId: id }}>
+      <NodeLink nodeId={id} nodeType="page">
         <div className="my-0.5 flex h-10 w-full cursor-pointer flex-row items-center gap-1 rounded-md p-1 hover:bg-accent">
           <Avatar size="small" id={id} name={name} avatar={avatar} />
           <div role="presentation" className="grow">
             {name}
           </div>
         </div>
-      </Link>
+      </NodeLink>
     </NodeViewWrapper>
   );
 };
